rifle_sxiv: properly fallback on non-valid extensions
originally we checked the extension to make sure the argument is an
image to avoid edge cases like this:

	./a        # is a dir
	./a.jpg

where running `rifle_sxiv a` would open up `a.jpg` instead of `a` the
directory.

however the is_img function was not correct as it would fail on upper
case extensions, such as `a.JPG`.

instead of doing extension check in two places, now we just have a
is_img_extension() function which checks for extension
case-insensitively via grep.

in short:

* this checks the extension case-insensitively
* reduces some code duplication
* sxiv will print an error msg for invalid files instead of the script
  silently exiting

this is probably a tad bit slower, but IMO worth the trade.

+++ examples/rifle_sxiv.sh
@@ -31,27 +31,27 @@
 TMPDIR="${TMPDIR:-/tmp}"
 tmp="$TMPDIR/sxiv_rifle_$$"
 
-listfiles () {
-    find -L "///${1%/*}" \( ! -path "///${1%/*}" -prune \) -type f -print |
-      grep -iE '\.(jpe?g|png|gif|svg|webp|tiff|heif|avif|ico|bmp)$' |
-      sort | tee "$tmp"
+is_img_extension () {
+    grep -iE '\.(jpe?g|png|gif|svg|webp|tiff|heif|avif|ico|bmp)$'
 }
 
-is_img () {
-    case "${1##*.}" in
-        "jpg"|"jpeg"|"png"|"gif"|"webp"|"tiff"|"bmp") return 0 ;;
-        *) return 1 ;;
-    esac
+listfiles () {
+    find -L "///${1%/*}" \( ! -path "///${1%/*}" -prune \) -type f -print |
+      is_img_extension | sort | tee "$tmp"
 }
 
 open_img () {
-    is_img "$1" || exit 1
-    trap 'rm -f $tmp' EXIT
-    count="$(listfiles "$1" | grep -nF "$1")"
+    # only go through listfiles() if the file has a valid img extension
+    if echo "$1" | is_img_extension >/dev/null 2>&1; then
+        trap 'rm -f $tmp' EXIT
+        count="$(listfiles "$1" | grep -nF "$1")"
+    fi
     if [ -n "$count" ]; then
         sxiv -i -n "${count%%:*}" -- < "$tmp"
     else
-        sxiv -- "$@" # fallback
+        # fallback incase file didn't have a valid extension, or we couldn't
+        # find it inside the list
+        sxiv -- "$@"
     fi
 }
